+++ klm/muid/src/MuidBuilder.cc
@@ -81,7 +81,7 @@ namespace Belle2 {
     }
 
     m_ReducedChiSquaredDx = MuidElementNumbers::getMaximalReducedChiSquared() /
-                            static_cast<double>(MuidElementNumbers::getSizeReducedChiSquared()); // bin size
+                            MuidElementNumbers::getSizeReducedChiSquared(); // bin size
     for (int detector = 0; detector <= MuidElementNumbers::getMaximalDetector(); ++detector) {
 
       for (int halfNdof = 1; halfNdof <= MuidElementNumbers::getMaximalHalfNdof(); ++halfNdof) {
